pyproject.toml
[tool.poetry]
name = "simple-ai-agents"
version = "0.1.0"
description = ""
authors = ["Timothy Lin <timothy0336@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fire = "^0.5.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
openai = "^0.28.1"
orjson = "^3.9.7"
huggingface-hub = "^0.17.3"
litellm = "^0.13.2"

[tool.poetry.group.dev.dependencies]
black = "^23.9.0"
pre-commit = "^2.20.0"
pytest = "^7.4.0"
rich = "^13.6.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
